Move formatting to new class AggregationFormatter
